requirements.txt
Flask
numpy
pandas
opencv-python-headless
tensorflow-cpu